Pulled grid dimensions into named constants
Following the principle that it's better not to have magic numbers
floating around in your code. Also, I need the game width and height to
be available so that I can keep the player within the bounds of the
game. And I'm thinking I'd like to make UnnamedGame use a Vec instead of
arrays, and I'll want that information as a variable for that too.