Added "attack" function
Right now all it does is remove a tile from the grid, so it's mostly a
placeholder
Also adds an IndexMut impl for Grid as part of that
I'm thinking I'm next going to stop using sprites for this game, at
least as part of the implementation of the game grid. I've been thinking
this for a while, but with the messing around with sprites I had to do
for this change, the thought has grown more immediate.